Changes premium store to reduce number of dependency cycle warnings
