+++ frontend/app/src/components/defi/DepositTotals.vue
@@ -58,7 +58,7 @@ import { useI18n } from 'vue-i18n-composable';
 import StatCardColumn from '@/components/display/StatCardColumn.vue';
 import StatCardWide from '@/components/display/StatCardWide.vue';
 import PremiumLock from '@/components/premium/PremiumLock.vue';
-import { getPremium } from '@/composables/premium';
+import { usePremium } from '@/composables/premium';
 
 defineProps({
   loading: {
@@ -78,7 +78,7 @@ defineProps({
     type: BigNumber
   }
 });
-const premium = getPremium();
+const premium = usePremium();
 
 const { t } = useI18n();
 </script>
