+++ frontend/app/src/components/profitloss/ProfitLossEvents.vue
@@ -136,7 +136,7 @@ import UpgradeRow from '@/components/history/UpgradeRow.vue';
 import CostBasisTable from '@/components/profitloss/CostBasisTable.vue';
 import ProfitLossEventType from '@/components/profitloss/ProfitLossEventType.vue';
 import { useRoute } from '@/composables/common';
-import { getPremium } from '@/composables/premium';
+import { usePremium } from '@/composables/premium';
 import {
   ProfitLossEvent,
   ProfitLossEvents,
@@ -265,7 +265,7 @@ const itemLength = computed(() => {
   return entriesFound;
 });
 
-const premium = getPremium();
+const premium = usePremium();
 
 const showUpgradeMessage = computed(
   () =>
