+++ frontend/app/src/composables/premium.ts
@@ -1,13 +1,7 @@
-import { set } from '@vueuse/core';
 import { storeToRefs } from 'pinia';
 import { usePremiumStore } from '@/store/session/premium';
 
-export const getPremium = () => {
+export const usePremium = () => {
   const { premium } = storeToRefs(usePremiumStore());
   return premium;
-};
-
-export const setPremium = (isPremium: boolean) => {
-  const { premium } = storeToRefs(usePremiumStore());
-  set(premium, isPremium);
 };
