+++ frontend/app/src/premium/setup-premium.ts
@@ -1,14 +1,12 @@
-import { set } from '@vueuse/core';
 import * as VueUse from '@vueuse/core';
 import * as BigNumber from 'bignumber.js';
 import * as Chart from 'chart.js';
 import ChartJsPluginZoom from 'chartjs-plugin-zoom';
-import { storeToRefs } from 'pinia';
 import * as Vue from 'vue';
+import { Ref } from 'vue';
 import * as zod from 'zod';
-import { usePremiumStore } from '@/store/session/premium';
 
-export const setupPremium = async () => {
+export const setupPremium = async (componentsLoaded: Ref<boolean>) => {
   window.Vue = Vue;
   window.VueUse = VueUse;
   window.Chart = Chart;
@@ -16,7 +14,6 @@ export const setupPremium = async () => {
   window.Chart = Chart;
   window.zod = zod;
   window.bn = BigNumber;
-  const { componentsLoaded } = storeToRefs(usePremiumStore());
   set(componentsLoaded, false);
   const { registerComponents } = await import('@/premium/register-components');
   registerComponents();
