+++ frontend/app/src/views/Statistics.vue
@@ -12,14 +12,14 @@
 import { defineComponent } from 'vue';
 import { useI18n } from 'vue-i18n-composable';
 import NoPremiumPlaceholder from '@/components/premium/NoPremiumPlaceholder.vue';
-import { getPremium } from '@/composables/premium';
+import { usePremium } from '@/composables/premium';
 import { PremiumStatistics } from '@/premium/premium';
 
 export default defineComponent({
   name: 'Statistics',
   components: { NoPremiumPlaceholder, PremiumStatistics },
   setup() {
-    const premium = getPremium();
+    const premium = usePremium();
     const { tc } = useI18n();
     return {
       premium,
